task management api completed

+++ api/src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import {TaskEntity} from "./task.entity";
 import {CreateTaskDto} from "./dto/create-task.dto";
 import {TaskStatus} from "./task-status.enum";
 import {GetTasksFilterDto} from "./dto/get-tasks-filter.dto";
+import {UserEntity} from "../auth/user.entity";
 
 @Injectable()
 export class TasksService {
@@ -12,12 +13,12 @@ export class TasksService {
     constructor(@InjectRepository(TaskRepository) private taskRepository: TaskRepository,) {
     }
 
-    async getAllTasks(filterDto: GetTasksFilterDto): Promise<TaskEntity[]> {
-        return await this.taskRepository.getTasks(filterDto);
+    async getAllTasks(filterDto: GetTasksFilterDto, user: UserEntity): Promise<TaskEntity[]> {
+        return await this.taskRepository.getTasks(filterDto, user);
     }
 
-    async getTaskById(id: number): Promise<TaskEntity> {
-        const task = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: UserEntity): Promise<TaskEntity> {
+        const task = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!task) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -25,20 +26,20 @@ export class TasksService {
         return task;
     }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<TaskEntity> {
-        return this.taskRepository.createTask(createTaskDto);
+    async createTask(createTaskDto: CreateTaskDto, user: UserEntity): Promise<TaskEntity> {
+        return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: UserEntity): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<TaskEntity> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: UserEntity): Promise<TaskEntity> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await task.save();
         return task;
